[Refactor] 카드 + 트래커 이미지 관련 내용 삭제 및 리팩토링
