+++ src/main/java/com/example/bookiibookii/domain/memberbook/exception/CardImageException.java
@@ -1,4 +1,4 @@
-package com.example.bookiibookii.domain.groupbook.exception;
+package com.example.bookiibookii.domain.memberbook.exception;
 
 import com.example.bookiibookii.global.apiPayload.code.BaseCode;
 import com.example.bookiibookii.global.apiPayload.exception.GeneralException;
