+++ src/main/java/com/example/bookiibookii/domain/memberbook/exception/code/CardImageErrorCode.java
@@ -1,4 +1,4 @@
-package com.example.bookiibookii.domain.groupbook.exception.code;
+package com.example.bookiibookii.domain.memberbook.exception.code;
 
 import com.example.bookiibookii.global.apiPayload.code.BaseCode;
 import lombok.Getter;
@@ -9,7 +9,6 @@ import org.springframework.http.HttpStatus;
 @RequiredArgsConstructor
 public enum CardImageErrorCode implements BaseCode {
 
-    //400 BAD_REQUEST
     INVALID_S3_KEY_FORMAT(HttpStatus.BAD_REQUEST, "CARDIMG400_1", "유효하지 않은 S3 키입니다. 형식: image/cards/{uuid}"),
     DUPLICATE_S3_KEY(HttpStatus.BAD_REQUEST, "CARDIMG400_2", "이미 사용 중인 S3 키입니다."),
     IMAGE_NOT_FOUND_IN_S3(HttpStatus.BAD_REQUEST, "CARDIMG400_3", "S3에 해당 이미지가 존재하지 않습니다."),
