+++ src/main/java/com/example/bookiibookii/domain/memberbook/service/CardImageValidationService.java
@@ -1,4 +1,4 @@
-package com.example.bookiibookii.domain.tracker.service;
+package com.example.bookiibookii.domain.memberbook.service;
 
 import org.springframework.stereotype.Component;
 
@@ -6,23 +6,29 @@ import java.util.UUID;
 import java.util.regex.Pattern;
 
 @Component
-public class TrackerImageValidationService {
+public class CardImageValidationService {
 
-    private static final Pattern S3_KEY_PATTERN = Pattern.compile("^image/trackers/([a-f0-9-]{36})$");
+    private static final Pattern S3_KEY_PATTERN = Pattern.compile("^image/cards/([a-f0-9-]{36})$");
 
     /**
-     * s3Key 형식 검증 (image/trackers/{uuid})
+     * s3Key 형식 검증 (UUID만 사용)
+     *
+     * @param s3Key 검증할 S3 키
+     * @return 검증 성공 여부
      */
     public boolean isValidS3Key(String s3Key) {
         if (s3Key == null || s3Key.isBlank()) {
             return false;
         }
+
         java.util.regex.Matcher matcher = S3_KEY_PATTERN.matcher(s3Key);
         if (!matcher.matches()) {
             return false;
         }
+
+        String uuidString = matcher.group(1);
         try {
-            UUID.fromString(matcher.group(1));
+            UUID.fromString(uuidString);
             return true;
         } catch (IllegalArgumentException e) {
             return false;
