+++ src/main/java/com/example/bookiibookii/domain/memberbook/service/MemberBookCardService.java
@@ -5,11 +5,9 @@ import com.example.bookiibookii.domain.group.exception.code.GroupErrorCode;
 import com.example.bookiibookii.domain.group.entity.MatchedMember;
 import com.example.bookiibookii.domain.group.repository.GroupsRepository;
 import com.example.bookiibookii.domain.group.repository.MatchedMemberRepository;
-import com.example.bookiibookii.domain.groupbook.dto.res.PresignedUrlResponseDTO;
-import com.example.bookiibookii.domain.groupbook.exception.CardImageException;
-import com.example.bookiibookii.domain.groupbook.exception.code.CardImageErrorCode;
-import com.example.bookiibookii.domain.groupbook.service.CardImageS3Service;
-import com.example.bookiibookii.domain.groupbook.service.CardImageValidationService;
+import com.example.bookiibookii.domain.memberbook.dto.res.PresignedUrlResponseDTO;
+import com.example.bookiibookii.domain.memberbook.exception.CardImageException;
+import com.example.bookiibookii.domain.memberbook.exception.code.CardImageErrorCode;
 import com.example.bookiibookii.domain.memberbook.dto.req.MemberCardCreateRequestDTO;
 import com.example.bookiibookii.domain.memberbook.dto.req.MemberCardUpdateRequestDTO;
 import com.example.bookiibookii.domain.memberbook.dto.res.MemberCardCreateResponseDTO;
