+++ src/main/java/com/example/bookiibookii/domain/tracker/dto/BookInfo.java
@@ -11,6 +11,5 @@ public record BookInfo(
 
         boolean isOwnerBook, // 현재 읽는 사람 기준으로 본인 소유 책인지
         String currentReaderNickname,
-        String currentReaderProfileImageUrl,
         int currentReadingRate
 ){}
