+++ src/main/java/com/example/bookiibookii/domain/tracker/dto/req/TrackerShippingRequestDTO.java
@@ -10,9 +10,5 @@ public record TrackerShippingRequestDTO(
 
         @Schema(description = "운송장 번호", example = "123456789012")
         @NotBlank(message = "송장 번호는 필수 입력입니다.")
-        String trackingNumber,
-
-        @Schema(description = "배송 인증 이미지 S3 키", example = "image/trackers/d6239b1e-dcd8-483c-baaf-018da2fa1328")
-        @NotBlank(message = "배송 인증 이미지 S3 키는 필수 입력입니다. Presigned URL로 업로드 후 발급받은 s3Key를 전달하세요.")
-        String s3Key
+        String trackingNumber
 ) {}
