--- src/main/java/com/example/bookiibookii/domain/tracker/dto/res/TrackerImageGetResponseDTO.java
@@ -1,11 +1,0 @@
-package com.example.bookiibookii.domain.tracker.dto.res;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class TrackerImageGetResponseDTO {
-    /** 인증 이미지 조회용 Presigned GET URL */
-    private String presignedGetUrl;
-}
