--- src/main/java/com/example/bookiibookii/domain/tracker/entity/TrackingImage.java
@@ -1,26 +1,0 @@
-package com.example.bookiibookii.domain.tracker.entity;
-
-import com.example.bookiibookii.global.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "tracking_image")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Builder
-public class TrackingImage extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "tracking_image_id")
-    private Long id;
-
-    @Column(name = "s3_key", length = 255, unique = true, nullable = false)
-    private String s3Key;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "delivery_id", nullable = false)
-    private Delivery delivery;
-}
